backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
pymysql==1.1.0
psycopg2-binary==2.9.9
redis==5.0.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# ML/Embeddings
sentence-transformers==2.3.1
numpy==1.26.3
faiss-cpu==1.7.4

# Vector DB
qdrant-client==1.7.3

# LLM
requests==2.31.0
httpx==0.26.0

# Utils
sqlparse==0.4.4
networkx==3.2.1
pandas==2.1.4

# Optional (for Kaggle)
kaggle==1.6.6
